requirements.txt
# Core dependencies for video analysis
numpy<2
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.14

# YOLO and deep learning
ultralytics
torch
torchvision

# Tracking and filtering
filterpy

# Web server (for backend_server.py and voice_coach_server.py)
flask
flask-cors
flask-socketio
python-dotenv

# Voice Coach - Speech Recognition (STT)
SpeechRecognition
pydub

# Voice Coach - Text-to-Speech (TTS)
elevenlabs

# Voice Coach - RAG/AI
pinecone-client
google-generativeai
anthropic

# Audio processing (required by pydub)
# Note: Also requires FFmpeg to be installed on system
# macOS: brew install ffmpeg
# Linux: sudo apt-get install ffmpeg
# Windows: Download from https://ffmpeg.org/download.html
